Realized permissions for car

+++ app/Http/Controllers/Api/Car/CarController.php
@@ -4,18 +4,20 @@ namespace App\Http\Controllers\Api\Car;
 
 use App\Models\Vehicle;
 use App\Services\CarService;
+use App\Services\PermissionService;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CreateCarRequest;
 use Illuminate\Support\Facades\Auth;
 
-
 class CarController extends Controller
 {
     private $carService;
+    private $permissionService;
 
-    public function __construct(CarService $carService)
+    public function __construct(CarService $carService, PermissionService $permissionService)
     {
         $this->carService = $carService;
+        $this->permissionService = $permissionService;
     }
 
     /**
@@ -25,11 +27,12 @@ class CarController extends Controller
      */
     public function index()
     {
-        return response()->json(
-            $this->carService->getAll()->map(function(Vehicle $cars){
-                return $cars;
-            })
-        );
+//        return response()->json(
+//            $this->carService->getAll()->map(function (Vehicle $cars) {
+//                return $cars;
+//            })
+//        );
+        return $this->carService->getAll();
     }
 
     /**
@@ -39,29 +42,39 @@ class CarController extends Controller
      */
     public function create()
     {
-        return true;
+        return $this->permissionService->canAddCar();
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @param CreateCarRequest $request
-     * @return Vehicle
+     * @return Vehicle|\Illuminate\Http\JsonResponse
      */
     public function store(CreateCarRequest $request)
     {
-        return $car = $this->carService->create($request);
+        if ($this->permissionService->canAddCar()) {
+            return $car = $this->carService->create($request);
+        } else {
+            return $this->accessDenied();
+        }
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
-        return $this->carService->getById($id);
+        $vehicle = $this->carService->getById($id);
+
+        if ($this->permissionService->canViewCar($vehicle->user_id)) {
+            return $vehicle;
+        } else {
+            return $this->accessDenied();
+        }
     }
 
     /**
@@ -72,7 +85,9 @@ class CarController extends Controller
      */
     public function edit($id)
     {
-        return $this->carService->getById($id);
+        $vehicle = $this->carService->getById($id);
+
+        return $this->permissionService->canEditCar($vehicle->id);
     }
 
     /**
@@ -84,7 +99,13 @@ class CarController extends Controller
      */
     public function update(CreateCarRequest $request, $id)
     {
-        return $this->carService->update($request, $id);
+        $vehicle = $this->carService->getById($id);
+
+        if ($this->permissionService->canEditCar($vehicle->id)) {
+            return $this->carService->update($request, $id);
+        } else {
+            return $this->accessDenied();
+        }
     }
 
     /**
@@ -95,6 +116,20 @@ class CarController extends Controller
      */
     public function destroy($id)
     {
-        return $this->carService->destroy($id);
+        $vehicle = $this->carService->getById($id);
+
+        if ($this->permissionService->canDeleteCar($vehicle->id)) {
+            return $this->carService->destroy($id);
+        } else {
+            return $this->accessDenied();
+        }
+    }
+
+    /**
+     * @return \Illuminate\Http\JsonResponse
+     */
+    private function accessDenied()
+    {
+        return response()->json('Access denied', 403);
     }
 }
